build fixes for elfutils-redhat on Android
For: AIA-452
Depends-Change-Id: Ied47b8059b046848437b870088a3fa4e08c9bccc

+++ libdw/libdwP.h
@@ -528,7 +528,7 @@ __libdw_in_section (Dwarf *dbg, int sec_index,
   if (data == NULL)
     return false;
   if (unlikely (addr < data->d_buf)
-      || unlikely (data->d_size - (addr - data->d_buf) < size))
+      || unlikely (data->d_size - ((size_t) addr - (size_t) data->d_buf) < size))
     {
       __libdw_seterrno (DWARF_E_INVALID_OFFSET);
       return false;
